feat: Add promise example JavaScript file
This commit adds a new JavaScript file called promise-example.js. The file contains a function called toastBread() which toasts bread for a given duration. It returns a promise that resolves if the duration is less than or equal to 3000 milliseconds and rejects with an error message if the duration is greater.

The promise is then used to demonstrate the different outcomes of toasting bread. If the duration is 4000 milliseconds, the promise will reject and an error message will be logged. Finally, a message indicating that the toaster is available is logged.